+++ CHANGELOG.md
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `df_to_dataset` task util for writing DataFrames to data lakes using `pyarrow`
 - Added retries to Cloud for Customers tasks
 - Added `chunksize` parameter to `C4CToDF` task to allow pulling data in chunks
+- Added `chunksize` parameter to `BCPTask` task to allow more control over the load process
 
 ### Changed
 - Changed the base class of `AzureSQL` to `SQLServer`